Create new user LMS
log in and create a new user LMS
